+++ src/fourth_bid/models/Product.java
@@ -7,7 +7,7 @@ public class Product {
     private int id;
     private String name;
     private String description;
-    private int provision;
+    private double provision;
     private int supplierID;
 
     private Supplier supplier;
@@ -28,7 +28,7 @@ public class Product {
         this.supplierID = supplierID;
     }
 
-    public Product(int id, String name, String description, int provision, int supplierID) {
+    public Product(int id, String name, String description, double provision, int supplierID) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -62,7 +62,7 @@ public class Product {
         this.description = description;
     }
 
-    public int getProvision() {
+    public double getProvision() {
         return provision;
     }
